Restructure the repo and refactor the template. Add necessary resources to host a static record set
