+++ .github/workflows/aws_sam.yml
@@ -3,6 +3,7 @@ on:
   push:
     branches:
       - main
+      - refactor-app-include-pulumi
   workflow_dispatch:
 
 jobs:
@@ -26,11 +27,13 @@ jobs:
       - name: Deploy to AWS Lambda
         env:
           AWS_ARN_ROLE: ${{ secrets.AWS_ARN_ROLE }}
-          AWS_SECRET_ID: ${{ secrets.AWS_SECRET_ID }}
           AWS_S3_BUCKET: ${{ secrets.AWS_S3_BUCKET }}
           OWNER: ${{ secrets.OWNER }}
           REPO: ${{ secrets.REPO }}
           WORKFLOW_FILE_NAME: ${{ secrets.WORKFLOW_FILE_NAME }}
           BRANCH: ${{ secrets.BRANCH }}
           JOB_FILTER: ${{ secrets.JOB_FILTER }}
-        run: make deploy-aws
+          HOSTED_ZONE_ID: ${{ secrets.HOSTED_ZONE_ID }}
+          FULL_DOMAIN_NAME: ${{ secrets.FULL_DOMAIN_NAME }}
+          TLS_CERTIFICATE_ARN: ${{ secrets.TLS_CERTIFICATE_ARN }}
+        run: make up
