+++ .gitignore
@@ -5,3 +5,4 @@ ghapp.pem
 ghrunner-app.json
 .gcloudignore
 .aws-sam
+dist
