+++ package.json
@@ -5,16 +5,23 @@
   "description": "Probot & Google Cloud Functions example",
   "scripts": {
     "start": "probot run ./app.js",
-    "test": "node test.js"
+    "test": "node test.js",
+    "build": "ncc build src/handler.js -o dist"
   },
   "author": "Gregor Martynus (https://twitter.com/gr2m)",
   "license": "ISC",
-  "repository": "github:probot/example-google-cloud-function",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/probot/example-google-cloud-function.git"
+  },
   "dependencies": {
     "@probot/adapter-aws-lambda-serverless": "^3.0.2",
+    "@pulumi/aws": "^5.34.0",
+    "@pulumi/pulumi": "^3.61.0",
     "probot": "^12.2.3"
   },
   "devDependencies": {
+    "@vercel/ncc": "^0.36.1",
     "nock": "^13.2.4",
     "smee-client": "^1.2.3",
     "uvu": "^0.5.3"
@@ -23,5 +30,14 @@
     "branches": [
       "main"
     ]
-  }
+  },
+  "bugs": {
+    "url": "https://github.com/probot/example-google-cloud-function/issues"
+  },
+  "homepage": "https://github.com/probot/example-google-cloud-function#readme",
+  "main": "index.js",
+  "directories": {
+    "test": "test"
+  },
+  "keywords": []
 }
