--- template.yml
@@ -1,56 +1,0 @@
-AWSTemplateFormatVersion: "2010-09-09"
-Transform: AWS::Serverless-2016-10-31
-Description: Serverless GitHub app
-
-Parameters:
-  awsRole:
-    Description: The AWS ARN role
-    Type: String
-  functionName:
-    Description: The AWS Lambda's Function name
-    Type: String
-  githubOwner:
-    Description: The repository owner
-    Type: String
-  githubRepository:
-    Description: The name of the repository
-    Type: String
-  githubWorkflowName:
-    Description: The workflow name
-    Type: String
-  githubBranch:
-    Description: The branch name
-    Type: String
-  githubJobFilter:
-    Description: The Job Filter
-    Type: String
-
-
-Resources:
-  webhooks:
-    Type: AWS::Serverless::Function
-    Properties:
-      FunctionName: !Ref functionName
-      Description: Basic Auth Funtion
-      CodeUri: .
-      Handler: handler.webhooks
-      Role: !Ref awsRole
-      Runtime: nodejs16.x
-      MemorySize: 256
-      Timeout: 20
-      PackageType: Zip
-      Tags:
-        map-migrated: d-server-01068mdjl5jze3
-      Events:
-        ApiEvent:
-          Type: HttpApi
-          Properties:
-            Path: /ghrunner-app
-            Method: post
-      Environment:
-        Variables:
-          OWNER: !Ref githubOwner
-          REPO: !Ref githubRepository
-          WORKFLOW_FILE_NAME: !Ref githubWorkflowName
-          BRANCH: !Ref githubBranch
-          JOB_FILTER: !Ref githubJobFilter
