+++ test/test.js
@@ -6,7 +6,7 @@ nock.disableNetConnect();
 
 const { Probot, ProbotOctokit } = require("probot");
 
-const app = require("./app");
+const app = require("../src/app");
 
 /** @type {import('probot').Probot */
 let probot;
